fixed) 현장 테스트 1. Defect Edit 오류 수정 2. Model Search 관련 오류 수정

+++ DefectDBManager/KoWeb_BCR/KoWeb/UDP/Packet.h
@@ -239,7 +239,7 @@ public:
 	void	MakeAckBcrSearchLotPacket(CString data, double progress);
 	
 	void	PullBcrSearchModelPatcket(char* buf, int buf_len, char* lotName);
-	void	MakeAckBcrSearchModelPacket(CString data, long progress);
+	void	MakeAckBcrSearchModelPacket(CString data, double progress);
 
 	void	PullBcrParamPacket(char* buf, int buf_len);
